Enable Label Sync GitHub Action
Adds an action to automatically sync labels used in this repository
from a centralized file in Project Mu DevOps.

https://github.com/microsoft/mu_devops

Any labels not specified in the label manifest located in Project Mu
DevOps will be removed from this repository when the action is run
and the labels will be kept in sync with those defined in the
manifest file.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>