MdeModulePkg: Enforce No Heap Guard on Certain Regions
This commit enforces the heap guard limitation that systems with
a runtime page allocation granularity that does not equal the
EFI_PAGE_SIZE cannot have page or pool guards set for
EfiRuntimeServicesCode, EfiRuntimeServicesData, EfiACPIMemoryNVS,
and EfiReservedMemoryType. This is to preserve alignment requirements
without extending the page guard size to very large granularities.

For each item, place an "x" in between `[` and `]` if true. Example: `[x]`.
_(you can also check items in the GitHub UI)_

- [x] Impacts functionality?
  - **Functionality** - Does the change ultimately impact how firmware functions?
  - Examples: Add a new library, publish a new PPI, update an algorithm, ...
- [x] Impacts security?
  - **Security** - Does the change have a direct security impact on an application,
    flow, or firmware?
  - Examples: Crypto algorithm change, buffer overflow fix, parameter
    validation improvement, ...
- [x] Breaking change?
  - **Breaking change** - Will anyone consuming this change experience a break
    in build or boot behavior?
  - Examples: Add a new library class, move a module to a different repo, call
    a function in a new library class in a pre-existing module, ...
- [ ] Includes tests?
  - **Tests** - Does the change include any explicit test code?
  - Examples: Unit tests, integration tests, robot tests, ...
- [ ] Includes documentation?
  - **Documentation** - Does the change contain explicit documentation additions
    outside direct code modifications (and comments)?
  - Examples: Update readme file, add feature readme file, link to documentation
    on an a separate Web page, ...

Tested on various physical and virtual platforms to confirm that AARCH64 cannot
set guards on these memory types and that x86 can.

For AARCH64 platforms (or in general platforms that do not have 4k runtime
page allocation granularity), do not set page or pool guards on
EfiRuntimeServicesCode, EfiRuntimeServicesData, EfiACPIMemoryNVS, or
EfiReservedMemoryType.