Set AARCH64 RUNTIME_PAGE_ALLOCATION_GRANULARITY to 64k instead of 4k
This reverts commit 6c80787f3d6b11aad287e8467c764a32b64c048d.

There were a lot of issues with 64k as a runtime page allocation
granularity for ARM64 when this was removed from Project Mu. These
issues have been fixed and 64k OSes have been confirmed to boot with
proper Memory Attribute Tables.

See https://github.com/microsoft/mu_basecore/pull/800 for the last
group of patches to update this.

For each item, place an "x" in between `[` and `]` if true. Example: `[x]`.
_(you can also check items in the GitHub UI)_

- [x] Impacts functionality?
  - **Functionality** - Does the change ultimately impact how firmware functions?
  - Examples: Add a new library, publish a new PPI, update an algorithm, ...
- [x] Impacts security?
  - **Security** - Does the change have a direct security impact on an application,
    flow, or firmware?
  - Examples: Crypto algorithm change, buffer overflow fix, parameter
    validation improvement, ...
- [x] Breaking change?
  - **Breaking change** - Will anyone consuming this change experience a break
    in build or boot behavior?
  - Examples: Add a new library class, move a module to a different repo, call
    a function in a new library class in a pre-existing module, ...
- [ ] Includes tests?
  - **Tests** - Does the change include any explicit test code?
  - Examples: Unit tests, integration tests, robot tests, ...
- [ ] Includes documentation?
  - **Documentation** - Does the change contain explicit documentation additions
    outside direct code modifications (and comments)?
  - Examples: Update readme file, add feature readme file, link to documentation
    on an a separate Web page, ...

Tested on various physical and virtual platforms to confirm MAT table correctly built on AARCH64 and AARCH64 64k page Linux can boot.

In the AARCH64 platform DSC, be sure to set:

```
[BuildOptions.common.EDKII.DXE_RUNTIME_DRIVER]
  MSFT:*_*_AARCH64_DLINK_FLAGS = /ALIGN:0x10000
  GCC:*_*_AARCH64_DLINK_FLAGS = -z common-page-size=0x10000
```

So that runtime images are aligned to 64k.